pm: seperate file to each function name
The pm_suspend, pm_resume and pm_staycount are in same file.
So, for code readability, Seperate function.
