+++ os/pm/Makefile
@@ -20,8 +20,10 @@
 
 ASRCS =
 
-CSRCS += pm_activity.c pm_changestate.c pm_checkstate.c pm_initialize.c pm_sleep.c
-CSRCS += pm_register.c pm_unregister.c pm_procfs.c pm_timedsuspend.c
+CSRCS += pm_changestate.c pm_checkstate.c pm_initialize.c
+CSRCS += pm_register.c pm_unregister.c pm_procfs.c
+CSRCS += pm_suspend.c pm_resume.c pm_timedsuspend.c pm_sleep.c
+CSRCS += pm_staycount.c
 
 ifeq ($(CONFIG_PM_METRICS),y)
 CSRCS += pm_metrics.c
